refactored configuration stuff to config file test cleanup
